rearrange imports and relocate functions
