+++ kobo_backup.py
@@ -1,40 +1,17 @@
+import argparse
+import datetime
+import glob
 import json
 import os
-import signal
-import datetime
 from pathlib import Path
 import platform
-import glob
+import shutil
+import signal
 import subprocess
 import sys
-import shutil
-import platform
-import argparse
 
-
-def get_directory_size(directory):  # figure out how much was backed up.
-    total = 0
-    try:
-        for entry in os.scandir(directory):
-            if entry.is_file():
-                total += entry.stat().st_size
-            elif entry.is_dir():
-                total += get_directory_size(entry.path)
-    except NotADirectoryError:
-        return os.path.getsize(directory)
-    except PermissionError:
-        return 0
-    return total
-
-
-def get_size_format(
-    b, factor=1024, suffix="B"
-):  # convert bytes to something human readable.
-    for unit in ["", "K", "M", "G", "T", "P", "E", "Z"]:
-        if b < factor:
-            return f"{b:.2f}{unit}{suffix}"
-        b /= factor
-    return f"{b:.2f}Y{suffix}"
+from automation_utils import create_linux_autostart_script
+from utils import get_directory_size, get_size_format
 
 
 def main(args):
@@ -69,7 +46,6 @@ def main(args):
     # Setup auto backup
     if args.auto:
         if platform.system() == "Linux":
-            from utils import create_linux_autostart_script
             # Create the autostart script
             create_linux_autostart_script()
             sys.exit()
