+++ linux_automation.py
@@ -1,9 +1,6 @@
 #!/usr/bin/env python3
-import gi
 from gi.repository import GLib, Gio
-import os
 import subprocess
-import sys
 
 class WatchForKobo(object):
